# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AI 智能下单与生产协同平台 API
  version: 1.0.0
  description: |
    牙科定制工厂三端系统（医生端 / 客服端 / 生产端 / 管理端）后端 API 规范。

    ## 认证方式
    所有接口（除登录外）必须携带 Authorization: Bearer {access_token}。
    服务端对每个接口强制校验角色权限，禁止仅靠前端隐藏。

    ## 角色枚举
    - DOCTOR：医生
    - CS：客服
    - WORKER：技工/生产人员
    - ADMIN：超级管理员

    ## 通用响应格式
    所有接口统一返回：
    { "code": 200, "msg": "success", "data": {} }
    错误时：
    { "code": 4xx/5xx, "msg": "错误说明", "data": null }
  license:
    name: Proprietary internal API
    url: https://yourdomain.com/legal/proprietary-api-license
servers:
- url: https://api.yourdomain.com/v1
  description: 正式环境
- url: https://test-api.yourdomain.com/v1
  description: 测试环境
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CommonResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        msg:
          type: string
          example: success
        data: {}
...
    ErrorResponse:
      type: object
      required:
      - code
      - msg
      - data
      properties:
        code:
          type: integer
          example: 403
        msg:
          type: string
          example: forbidden
        data:
          nullable: true
          type: object
...
    AiOrderOnlyRequest:
      type: object
      required:
      - order_id
      properties:
        order_id:
          type: integer
    AiQuestionRequest:
      type: object
      required:
      - order_id
      - question
      properties:
        order_id:
          type: integer
        question:
          type: string
    AiTranslateRequest:
      type: object
      required:
      - order_id
      - source_text
      properties:
        order_id:
          type: integer
        source_text:
          type: string
    AiTranslateResponse:
      type: object
      properties:
        translated_text:
          type: string
    AiAnswerResponse:
      type: object
      properties:
        answer:
          type: string
        reference_data_notes:
          type: array
          description: AI-2 客服查询使用的只读内部数据来源说明；其他 AI 回答可为空。
          items:
            type: string
    AiProductionNoteResponse:
      type: object
      properties:
        draft_note:
          type: string
        template_version:
          type: string
          description: AI-5 生产备注默认模板版本；当前为 PHASE_ONE_DEFAULT_V1。
        knowledge_context_notes:
          type: array
          description: AI-5 草稿使用的知识上下文说明，包含订单、表单、消息、附件、账单、物流等只读来源。
          items:
            type: string
        requires_customer_template_confirmation:
          type: boolean
          description: true 表示客户正式模板仍待客户 / PM 确认，不能声明为真实客户模板。
    AiProductionNoteConfirmRequest:
      type: object
      required:
      - order_id
      - draft_note
      properties:
        order_id:
          type: integer
        draft_note:
          type: string
          description: 已由 CS / WORKER / ADMIN 人工确认的 AI-5 生产备注草稿。
        confirmation_note:
          type: string
          description: 人工确认说明，可为空。
    AiProductionNoteConfirmResponse:
      type: object
      properties:
        production_note:
          type: string
          description: 追加写入后的订单内部生产备注。
        template_version:
          type: string
          description: AI-5 生产备注默认模板版本；当前为 PHASE_ONE_DEFAULT_V1。
        requires_customer_template_confirmation:
          type: boolean
          description: true 表示客户正式模板仍待客户 / PM 确认。
    AiGovernanceCostTrendResponse:
      type: object
      properties:
        days:
          type: integer
          description: 查询窗口天数，默认 7，服务端限制为 1 到 31。
        points:
          type: array
          items:
            "$ref": "#/components/schemas/AiGovernanceCostTrendPoint"
        total_success_count:
          type: integer
          format: int64
        total_estimated_cost_microusd:
          type: integer
          format: int64
    AiGovernanceCostTrendPoint:
      type: object
      properties:
        date:
          type: string
          format: date
          description: 聚合日期。
        success_count:
          type: integer
          format: int64
          description: 当天成功模型调用次数。
        estimated_cost_microusd:
          type: integer
          format: int64
          description: 当天成功模型调用估算成本，单位 microUSD。
        model_count:
          type: integer
          format: int64
          description: 当天出现过成功调用的模型数量。
    AiExternalAlertSummaryResponse:
      type: object
      properties:
        status_counts:
          type: array
          description: 按 send_status 聚合的 outbox 数量分布。
          items:
            "$ref": "#/components/schemas/AiExternalAlertStatusCount"
        pending_count:
          type: integer
          format: int64
        sending_count:
          type: integer
          format: int64
        sent_count:
          type: integer
          format: int64
        failed_count:
          type: integer
          format: int64
        dead_letter_count:
          type: integer
          format: int64
        latest_failure:
          nullable: true
          "$ref": "#/components/schemas/AiExternalAlertFailure"
        oldest_pending_created_at:
          type: string
          format: date-time
          nullable: true
          description: 最老一条 PENDING outbox 的创建时间；无待发送记录时为空。
    AiExternalAlertStatusCount:
      type: object
      properties:
        send_status:
          type: string
          enum:
          - PENDING
          - SENDING
          - SENT
          - FAILED
          - DEAD_LETTER
        count:
          type: integer
          format: int64
    AiExternalAlertFailure:
      type: object
      properties:
        alert_id:
          type: integer
          format: int64
        alert_type:
          type: string
        send_status:
          type: string
          enum:
          - FAILED
          - DEAD_LETTER
        attempts:
          type: integer
        last_error:
          type: string
          nullable: true
          description: 已做基础脱敏的最近错误摘要，不返回真实 webhook URL、密钥、prompt 原文或模型原始响应。
        updated_at:
          type: string
          format: date-time
    AiExternalAlertListResponse:
      type: object
      properties:
        limit:
          type: integer
          description: 本次查询使用的返回上限，服务端限制为 1 到 100。
        records:
          type: array
          items:
            "$ref": "#/components/schemas/AiExternalAlertRecord"
    AiExternalAlertRecord:
      type: object
      properties:
        alert_id:
          type: integer
          format: int64
        event_type:
          type: string
          description: outbox 的 alert_type，对外按 event_type 呈现。
        send_status:
          type: string
          enum:
          - PENDING
          - SENDING
          - SENT
          - FAILED
          - DEAD_LETTER
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        attempts:
          type: integer
          description: outbox 当前累计发送尝试次数。
        last_error:
          type: string
          nullable: true
          description: 仅 FAILED / DEAD_LETTER 记录返回已脱敏错误摘要；不返回真实 webhook URL、密钥、prompt 原文、模型原始响应或上游敏感响应。
        last_attempted_at:
          type: string
          format: date-time
          nullable: true
          description: 最近一次发送尝试时间；无尝试记录时为 null。
    AiExternalAlertReceiverResponse:
      type: object
      properties:
        accepted:
          type: boolean
          description: 当前本地接收端验签是否通过。
        event_type:
          type: string
          description: 从请求 JSON 的 event 字段提取的事件类型；不会返回原始 payload。
        nonce:
          type: string
          description: 已接受的 nonce；同一进程窗口内重复 nonce 会被拒绝。
    MissingInfoResponse:
      type: object
      properties:
        is_complete:
          type: boolean
        missing_items:
          type: array
          items:
            type: object
            properties:
              field_key:
                type: string
              field_label:
                type: string
              tip:
                type: string
...
  responses:
    BadRequest:
      description: 请求参数错误或校验失败
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: 未登录或 Token 无效
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    Forbidden:
      description: 无权限或数据范围拒绝
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    Conflict:
      description: 业务状态冲突或幂等冲突
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    ServiceUnavailable:
      description: 外部依赖或存储服务不可用
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    DefaultError:
      description: 未分类错误响应
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
security:
- BearerAuth: []
paths:
...
  "/ai/translate":
    post:
      tags:
      - AI
      summary: AI-1 翻译助手（客服端）
      description: |
        权限：CS / ADMIN
        返回翻译草稿，客服确认后需调用 /orders/{orderId}/review 写入生产指令。
        本接口只负责生成草稿，不写入任何订单字段。
        任务 9D.15 第一增量：当后端配置 AI_PROVIDER=deepseek、AI_DEEPSEEK_ENABLED=true 且注入 DEEPSEEK_API_KEY 时，通过 OpenAI-compatible /chat/completions 调用 DeepSeek；未启用或未配置真实 key 时保持 deterministic-placeholder 安全占位输出。
        任务 9D.35 第一增量：当 AI_BUDGET_CIRCUIT_BREAKER_ENABLED=true 且近 24 小时成功调用估算成本已达到 AI_DAILY_BUDGET_MICROUSD 时，本接口不再外呼真实模型，返回 deterministic fallback，并写入 AI_BUDGET_CIRCUIT_OPEN 治理审计。
        任务 9D.37 第一增量：预算跨线 AI_BUDGET_EXCEEDED 或预算熔断 AI_BUDGET_CIRCUIT_OPEN 会写入 ai_external_alert_outbox 待发送事实，send_status=PENDING；当前不调用真实外部告警渠道。
        任务 9D.38 第一增量：当 AI_BUDGET_CIRCUIT_BREAKER_ENABLED=true 且当前角色的 AI_ADMIN_DAILY_BUDGET_MICROUSD / AI_CS_DAILY_BUDGET_MICROUSD / AI_DOCTOR_DAILY_BUDGET_MICROUSD / AI_WORKER_DAILY_BUDGET_MICROUSD 已超限时，不再外呼真实模型，并写入 AI_BUDGET_ROLE_CIRCUIT_OPEN 治理审计和 outbox。
        任务 9D.39 第一增量：当 AI_BUDGET_CIRCUIT_BREAKER_ENABLED=true 且当前 AI_DEEPSEEK_MODEL 的 AI_DEEPSEEK_DAILY_BUDGET_MICROUSD 已超限时，不再外呼真实模型，并写入 AI_BUDGET_MODEL_CIRCUIT_OPEN 治理审计和 outbox。
        任务 9D.40 第一增量：AI 审计写入 prompt_version；真实模型输出命中敏感防护时，返回安全保护文案，并写入 AI_OUTPUT_GUARDED / ai-governance-output-guard 治理审计。
        任务 9D.41 第一增量：ai_external_alert_outbox 的 PENDING 记录可由本地发送器推进为 SENT 或 FAILED，并累计 attempts / 记录 last_error。
        任务 9D.43 第一增量：发送器默认仍是本地 dry-run；只有显式配置 AI_EXTERNAL_ALERT_WEBHOOK_ENABLED=true 且提供 AI_EXTERNAL_ALERT_WEBHOOK_URL 时，EXTERNAL_ALERT 通道才会把 outbox payload 以 application/json POST 到外部 webhook，不提交任何真实渠道密钥。
        任务 9D.44 第一增量：AI 外部告警调度器默认关闭；只有显式配置 AI_EXTERNAL_ALERT_SCHEDULER_ENABLED=true 时，才会按 AI_EXTERNAL_ALERT_SCHEDULER_BATCH_SIZE 批量调用发送器处理 PENDING outbox。本增量不做分布式锁、复杂重试、死信或真实渠道密钥。
        任务 9D.45 第一增量：webhook 非 2xx 或连接异常会累计 attempts 并写 last_error；未达到 AI_EXTERNAL_ALERT_MAX_ATTEMPTS 时保持 PENDING 等待下次调度，达到上限后标记 DEAD_LETTER。本增量不做分布式锁、退避调度、死信管理页面或生产 webhook 联调。
        任务 9D.46 第一增量：发送器会先用条件更新把 PENDING outbox 领取为事务内 SENDING，领取成功后才允许外呼 webhook；重复触发或并发 sender 不能对同一条 outbox 造成重复外呼。本增量不做签名/鉴权、退避调度、告警抑制、监控指标或生产 webhook 联调。
        任务 9D.47 第一增量：默认不签名；显式设置 AI_EXTERNAL_ALERT_WEBHOOK_SIGNING_ENABLED=true 并通过安全渠道注入 AI_EXTERNAL_ALERT_WEBHOOK_SIGNING_SECRET 后，webhook 请求会携带 X-AI-Alert-Timestamp、X-AI-Alert-Nonce 和 X-AI-Alert-Signature: sha256=<HMAC-SHA256(timestamp.nonce.requestBody)>。本增量不提交真实 secret，不做生产 webhook 联调。
        任务 9D.71 第一增量：新增本地接收端验签 / 防重放验收桩 `/ai/external-alerts/receive`，默认 `AI_EXTERNAL_ALERT_RECEIVER_VERIFICATION_ENABLED=false`；显式启用并注入 `AI_EXTERNAL_ALERT_RECEIVER_SIGNING_SECRET` 后，会校验 timestamp 时间窗、nonce 重放和 HMAC 签名。本增量不接真实外部服务，不提交真实 secret。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AiTranslateRequest"
      responses:
        '200':
          description: 返回翻译草稿，不自动写入业务字段
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/AiTranslateResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: postAiTranslate
  "/ai/check-missing":
    post:
      tags:
      - AI
      summary: AI-4 资料缺失检查助手（医生端 + 客服端）
      description: |
        权限：DOCTOR / CS / ADMIN
        只返回缺失项清单，不执行任何驳回或退回操作。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AiOrderOnlyRequest"
      responses:
        '200':
          description: 返回缺失资料检查结果
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/MissingInfoResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: postAiCheckMissing
  "/ai/cs-query":
    post:
      tags:
      - AI
      summary: AI-2 客服查询助手（客服端）
      description: |
        权限：CS / ADMIN
        读取客服权限范围内的全部内部数据。
        返回自然语言回答，对外发送前客服必须确认。
        本接口不写入任何数据，不发送任何消息。
        任务 9D.15 第一增量：可通过后端 DeepSeek 适配层生成草稿；所有调用写 ai_audit_log.model_name，未启用真实模型时保持 deterministic-placeholder。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AiQuestionRequest"
      responses:
        '200':
          description: 返回客服内部查询草稿
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/AiAnswerResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: postAiCsQuery
...
  "/ai/production-note":
    post:
      tags:
      - AI
      summary: AI-5 生产备注助手（客服端 + 生产端）
      description: |
        权限：CS / WORKER / ADMIN
        返回生产备注草稿和知识上下文说明。客户正式模板仍未确认，当前使用 PHASE_ONE_DEFAULT_V1 默认模板并要求人工确认。
        本接口不写入任何数据，草稿需通过 /ai/production-note/confirm 人工确认后才追加保存到订单内部 production_note。
        任务 9D.15 第一增量：可通过后端 DeepSeek 适配层生成草稿；本接口仍不自动写入订单字段。
        任务 9D.98 第一增量：补默认模板版本、知识上下文说明和人工确认写入链路。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AiOrderOnlyRequest"
      responses:
        '200':
          description: 返回生产备注草稿，不自动写订单字段
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/AiProductionNoteResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: postAiProductionNote
  "/ai/production-note/confirm":
    post:
      tags:
      - AI
      summary: AI-5 生产备注人工确认写入（客服端 + 生产端）
      description: |
        权限：CS / WORKER / ADMIN
        将人工确认后的 AI-5 生产备注草稿追加写入订单内部 production_note，并写入 AI 审计。
        该接口不代表客户正式模板已确认；requires_customer_template_confirmation=true 时仍需客户 / PM 最终确认。
        不接真实 DeepSeek key、不自动下发生产指令、不自动发送给医生。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AiProductionNoteConfirmRequest"
      responses:
        '200':
          description: 返回写入后的内部生产备注
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/AiProductionNoteConfirmResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: postAiProductionNoteConfirm
  "/ai/governance/cost-trend":
    get:
      tags:
      - AI
      summary: AI 成本趋势
      description: |
        权限：CS / ADMIN
        任务 9D.42 第一增量：按天聚合 ai_audit_log 中 result_status=SUCCESS 的成功模型调用，返回成功次数、估算成本 microUSD 和当日模型数量，用于上线前观察 AI 成本变化。
        当前只读取已有审计表，不接真实计费账单、不做预算策略管理页面、不导出报表，也不向医生端/生产端开放。
      parameters:
      - name: days
        in: query
        required: false
        schema:
          type: integer
          default: 7
          minimum: 1
          maximum: 31
        description: 查询最近多少天；服务端会限制到 1 到 31 天。
      responses:
        '200':
          description: 最近 N 天 AI 成本趋势
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/AiGovernanceCostTrendResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: getAiGovernanceCostTrend
  "/ai/governance/external-alerts/summary":
    get:
      tags:
      - AI
      summary: AI 外部告警 outbox 监控摘要
      description: |
        权限：CS / ADMIN
        任务 9D.48 第一增量：只读聚合 `ai_external_alert_outbox` 运维可观察信息，返回 `PENDING/SENDING/SENT/FAILED/DEAD_LETTER` 数量分布、最近一条 FAILED/DEAD_LETTER 错误和最老 PENDING 创建时间。
        本接口不做真实 webhook 联调，不发送短信/邮件/企业微信，不返回真实 webhook URL、密钥、prompt 原文、模型原始响应或内部生产敏感详情，也不提供人工重放、关闭、抑制等运维动作。
      responses:
        '200':
          description: AI 外部告警 outbox 监控摘要
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/AiExternalAlertSummaryResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: getAiExternalAlertSummary
  "/ai/governance/external-alerts":
    get:
      tags:
      - AI
      summary: AI 外部告警 outbox 列表
      description: |
        权限：CS / ADMIN
        任务 9D.48.1 第一增量：只读查看 `ai_external_alert_outbox` 最近记录列表，支持 `send_status`、`event_type`、`created_at` 起止范围和 `limit` 最小筛选。
        任务 9D.48.2 第一增量：FAILED / DEAD_LETTER 记录额外返回 `attempts`、已脱敏 `last_error` 和 `last_attempted_at`，用于只读排查失败/死信原因。
        响应不返回密钥、真实 webhook URL、payload、prompt 原文、模型原始响应、上游敏感响应或内部生产敏感详情；本接口不提供人工重放、编辑、死信恢复或人工处理状态。
      parameters:
      - name: send_status
        in: query
        required: false
        schema:
          type: string
          enum:
          - PENDING
          - SENDING
          - SENT
          - FAILED
          - DEAD_LETTER
        description: 按 outbox send_status 精确筛选。
      - name: event_type
        in: query
        required: false
        schema:
          type: string
        description: 按 outbox alert_type 精确筛选。
      - name: created_at_from
        in: query
        required: false
        schema:
          type: string
          format: date-time
        description: ISO-8601 本地时间下限，例如 2026-07-04T00:00:00。
      - name: created_at_to
        in: query
        required: false
        schema:
          type: string
          format: date-time
        description: ISO-8601 本地时间上限，例如 2026-07-04T23:59:59。
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 20
          minimum: 1
          maximum: 100
        description: 返回上限；服务端限制为 1 到 100。
      responses:
        '200':
          description: AI 外部告警 outbox 最近记录
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/AiExternalAlertListResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: getAiExternalAlerts
  "/ai/external-alerts/receive":
    post:
      tags:
      - AI
      summary: AI 外部告警接收端验签 / 防重放验收桩
      description: |
        任务 9D.71 第一增量：本接口用于本地和测试环境验证外部告警接收端的 timestamp / nonce / HMAC 签名口径。
        默认 `AI_EXTERNAL_ALERT_RECEIVER_VERIFICATION_ENABLED=false`，接口返回 503，不接收任何真实外部服务。
        显式启用后，调用方必须提供 `X-AI-Alert-Timestamp`、`X-AI-Alert-Nonce` 和 `X-AI-Alert-Signature`；签名基串为 `timestamp.nonce.requestBody`，签名值为 `sha256=<HMAC-SHA256>`。
        同一进程 replay window 内重复 nonce 返回 409；超出 `AI_EXTERNAL_ALERT_RECEIVER_REPLAY_WINDOW_SECONDS` 时间窗或签名不匹配返回 401。
        响应不返回原始 payload、真实 webhook URL、密钥、prompt 原文、模型原始响应或内部生产敏感详情。
      parameters:
      - name: X-AI-Alert-Timestamp
        in: header
        required: true
        schema:
          type: string
        description: Unix epoch seconds；用于时间窗校验。
      - name: X-AI-Alert-Nonce
        in: header
        required: true
        schema:
          type: string
        description: 调用方生成的一次性随机值；同一接收端窗口内不可重复。
      - name: X-AI-Alert-Signature
        in: header
        required: true
        schema:
          type: string
        description: sha256=<HMAC-SHA256(timestamp.nonce.requestBody)>。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: 接收端验签通过
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/AiExternalAlertReceiverResponse"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
        default:
          "$ref": "#/components/responses/DefaultError"
      operationId: postAiExternalAlertReceive
...